Separate additional classes out of 'RegistrationData'
