+++ app/src/main/java/com/rockthevote/grommet/ui/registration/InvalidRegistrationException.kt
@@ -1,0 +1,9 @@
+package com.rockthevote.grommet.ui.registration
+
+/**
+ * Provided registration data is invalid.
+ */
+class InvalidRegistrationException(
+    msg: String,
+    val userMessage: String
+) : Exception(msg)
